Proper PDO mapping in the SDO and right type for the uint16_t

+++ include/iit/ecat/advr/mc_lowpwr_esc.h
@@ -320,7 +320,7 @@ public:
 
         std::ostringstream oss;
         float act_position;
-        int32_t fault;
+        uint16_t fault;
         //uint16_t gain;
         float gain;
         float test_ref;
